refactor: update profile page
