+++ src/app/components/profile/sections/profile-form/profile-form.component.html
@@ -1,0 +1,92 @@
+<div class="space-y-6">
+  <div class="border-b border-gray-200 pb-4">
+    <label
+      for="user-name"
+      class="block text-sm font-medium text-gray-600 mb-2">
+      Nome
+    </label>
+    <p *ngIf="!isEditing" class="text-lg text-gray-800">{{ userName || 'Não disponível' }}</p>
+    <input
+      *ngIf="isEditing"
+      id="user-name"
+      type="text"
+      [(ngModel)]="editedName"
+      required
+      minlength="2"
+      autocomplete="name"
+      class="w-full px-4 py-2 border border-gray-300 rounded-lg text-lg text-gray-800 focus:outline-none focus:ring-2 focus:ring-primary-500 focus:border-transparent transition-colors"
+      [class.border-red-500]="profileState.errorMessage && !editedName.trim()"
+      placeholder="Digite seu nome"
+      aria-label="Nome do usuário"
+      aria-required="true">
+  </div>
+
+  <div class="border-b border-gray-200 pb-4">
+    <label
+      for="user-email"
+      class="block text-sm font-medium text-gray-600 mb-2">
+      Email
+    </label>
+    <p *ngIf="!isEditing" class="text-lg text-gray-800">{{ userEmail || 'Não disponível' }}</p>
+    <input
+      *ngIf="isEditing"
+      id="user-email"
+      type="email"
+      [(ngModel)]="editedEmail"
+      required
+      autocomplete="email"
+      class="w-full px-4 py-2 border border-gray-300 rounded-lg text-lg text-gray-800 focus:outline-none focus:ring-2 focus:ring-primary-500 focus:border-transparent transition-colors"
+      [class.border-red-500]="profileState.errorMessage && !editedEmail.trim()"
+      placeholder="Digite seu email"
+      aria-label="Email do usuário"
+      aria-required="true">
+  </div>
+
+  <!-- Slot: conteúdo entre campos do perfil e botões (ex.: senha) -->
+  <ng-content select="[between]"></ng-content>
+
+  <div class="flex flex-col sm:flex-row gap-4 mt-8">
+    <button
+      *ngIf="!isEditing"
+      (click)="startEditing()"
+      type="button"
+      class="flex-1 py-3 bg-gradient-to-br from-primary-500 to-primary-700 text-white border-none rounded-lg text-base font-semibold cursor-pointer transition-all duration-200 hover:-translate-y-0.5 hover:shadow-[0_5px_15px_rgba(196,235,110,0.4)] focus:outline-none focus:ring-2 focus:ring-primary-500 focus:ring-offset-2"
+      aria-label="Editar informações do perfil">
+      Editar Perfil
+    </button>
+
+    <button
+      *ngIf="isEditing"
+      (click)="cancelEditing()"
+      type="button"
+      [disabled]="isLoading"
+      class="flex-1 py-3 bg-gray-500 text-white border-none rounded-lg text-base font-semibold cursor-pointer transition-all duration-200 hover:-translate-y-0.5 hover:shadow-[0_5px_15px_rgba(107,114,128,0.4)] disabled:opacity-50 disabled:cursor-not-allowed focus:outline-none focus:ring-2 focus:ring-gray-500 focus:ring-offset-2"
+      aria-label="Cancelar edição">
+      Cancelar
+    </button>
+
+    <button
+      *ngIf="isEditing"
+      (click)="saveProfile()"
+      type="button"
+      [disabled]="isLoading"
+      class="flex-1 py-3 bg-gradient-to-br from-primary-500 to-primary-700 text-white border-none rounded-lg text-base font-semibold cursor-pointer transition-all duration-200 hover:-translate-y-0.5 hover:shadow-[0_5px_15px_rgba(196,235,110,0.4)] disabled:opacity-50 disabled:cursor-not-allowed focus:outline-none focus:ring-2 focus:ring-primary-500 focus:ring-offset-2"
+      aria-label="Salvar alterações do perfil"
+      [attr.aria-busy]="isLoading">
+      <span *ngIf="!isLoading">Salvar</span>
+      <span *ngIf="isLoading" class="flex items-center justify-center gap-2">
+        <span class="inline-block w-4 h-4 border-2 border-white border-t-transparent rounded-full animate-spin"></span>
+        Salvando...
+      </span>
+    </button>
+
+    <button
+      (click)="logout()"
+      type="button"
+      class="flex-1 py-3 bg-gradient-to-br from-red-500 to-pink-600 text-white border-none rounded-lg text-base font-semibold cursor-pointer transition-all duration-200 hover:-translate-y-0.5 hover:shadow-[0_5px_15px_rgba(239,68,68,0.4)] focus:outline-none focus:ring-2 focus:ring-red-500 focus:ring-offset-2"
+      aria-label="Sair da conta">
+      Sair
+    </button>
+  </div>
+</div>
+
