+++ src/app/components/profile/sections/profile-image/profile-image-section.component.html
@@ -1,0 +1,83 @@
+<!-- Foto de perfil -->
+<div class="border-b border-gray-200 pb-6">
+  <div class="flex items-start justify-between gap-4 mb-4">
+    <div>
+      <h2 class="text-xl font-semibold text-gray-800">Foto do Perfil</h2>
+      <p class="text-sm text-gray-500 mt-1">
+        Envie uma imagem (PNG/JPG/WEBP) de até 2MB.
+      </p>
+    </div>
+    <span class="inline-flex items-center gap-2 px-3 py-1 rounded-full bg-gray-100 text-gray-600 text-xs font-medium border border-gray-200">
+      <span class="inline-block w-2 h-2 rounded-full bg-primary-500"></span>
+      Perfil
+    </span>
+  </div>
+
+  <div class="flex flex-col sm:flex-row items-start sm:items-center gap-4">
+    <!-- input escondido: aberto ao clicar no avatar -->
+    <input
+      #profileImageInput
+      type="file"
+      accept="image/png,image/jpeg,image/webp"
+      (change)="onSelectProfileImage($event)"
+      [disabled]="isUploadingProfileImage"
+      class="hidden" />
+
+    <button
+      type="button"
+      class="relative group"
+      (click)="!isUploadingProfileImage && profileImageInput.click()"
+      (keydown.enter)="!isUploadingProfileImage && profileImageInput.click()"
+      (keydown.space)="!isUploadingProfileImage && profileImageInput.click()"
+      [disabled]="isUploadingProfileImage"
+      aria-label="Alterar foto do perfil">
+      <div class="w-20 h-20 rounded-full overflow-hidden bg-gray-100 border border-gray-200 flex items-center justify-center group-hover:border-gray-400 transition-colors">
+        <img
+          *ngIf="displayedAvatarUrl; else avatarFallback"
+          [src]="displayedAvatarUrl"
+          alt="Foto do perfil"
+          class="w-full h-full object-cover" />
+        <ng-template #avatarFallback>
+          <span class="text-xl font-semibold text-gray-600">{{ profileInitials }}</span>
+        </ng-template>
+      </div>
+      <div class="absolute inset-0 rounded-full bg-black/0 group-hover:bg-black/10 transition-colors"></div>
+      <div *ngIf="profileImagePreviewUrl" class="absolute -bottom-2 -right-2 px-2 py-1 rounded-full bg-yellow-100 text-yellow-800 text-[10px] font-semibold border border-yellow-200">
+        Preview
+      </div>
+    </button>
+
+    <div class="flex-1 w-full">
+      <div *ngIf="profileImageSuccessMessage" class="mb-3 p-3 bg-green-100 border border-green-400 text-green-700 rounded-lg">
+        {{ profileImageSuccessMessage }}
+      </div>
+      <div *ngIf="profileImageErrorMessage" class="mb-3 p-3 bg-red-100 border border-red-400 text-red-700 rounded-lg">
+        {{ profileImageErrorMessage }}
+      </div>
+
+      <div class="flex gap-3">
+        <button
+          type="button"
+          (click)="cancelProfileImageSelection(profileImageInput)"
+          [disabled]="isUploadingProfileImage && !profileImagePreviewUrl"
+          class="py-3 px-4 bg-white text-gray-700 border border-gray-300 rounded-lg text-sm font-semibold cursor-pointer transition-all duration-200 hover:bg-gray-100 disabled:opacity-50 disabled:cursor-not-allowed focus:outline-none focus:ring-2 focus:ring-gray-500 focus:ring-offset-2">
+          Limpar
+        </button>
+
+        <button
+          type="button"
+          (click)="uploadProfileImage()"
+          [disabled]="!selectedProfileImageFile || isUploadingProfileImage"
+          class="py-3 px-4 bg-gradient-to-br from-primary-500 to-primary-700 text-white border-none rounded-lg text-sm font-semibold cursor-pointer transition-all duration-200 hover:-translate-y-0.5 hover:shadow-[0_5px_15px_rgba(196,235,110,0.35)] disabled:opacity-50 disabled:cursor-not-allowed focus:outline-none focus:ring-2 focus:ring-primary-500 focus:ring-offset-2"
+          [attr.aria-busy]="isUploadingProfileImage">
+          <span *ngIf="!isUploadingProfileImage">Salvar Foto</span>
+          <span *ngIf="isUploadingProfileImage" class="flex items-center justify-center gap-2">
+            <span class="inline-block w-4 h-4 border-2 border-white border-t-transparent rounded-full animate-spin"></span>
+            Enviando...
+          </span>
+        </button>
+      </div>
+    </div>
+  </div>
+</div>
+
