+++ src/app/components/profile/shared/profile-http.utils.ts
@@ -1,0 +1,4 @@
+export function getServerMessage(error: any): string | null {
+  return error?.error?.message || error?.error?.error || null;
+}
+
